[FSSS-172] Fix/Adjust inappropriate rerenders (#304)
