+++ src/components/ui/Breadcrumb/index.tsx
@@ -1,2 +1,3 @@
 export { default } from './Breadcrumb'
+export { default as BreadcrumbWrapper } from './BreadcrumbWrapper'
 export type { BreadcrumbProps } from './Breadcrumb'
